+++ src/main/java/hello/QueryController.java
@@ -9,22 +9,22 @@ import org.springframework.web.bind.annotation.RestController;
 public class QueryController {
 
     @RequestMapping("/add")
-    public Addition addition(@RequestParam(value="num1", defaultValue="5") int num1, @RequestParam(value="num2", defaultValue="5") int num2) {
+    public Addition addition(@RequestParam(value="num1", defaultValue="5") float num1, @RequestParam(value="num2", defaultValue="5") float num2) {
         return new Addition(num1, num2);
     }
 	
 	@RequestMapping("/minus")
-    public Subtraction subtraction(@RequestParam(value="num1", defaultValue="5") int num1, @RequestParam(value="num2", defaultValue="5") int num2) {
+    public Subtraction subtraction(@RequestParam(value="num1", defaultValue="5") float num1, @RequestParam(value="num2", defaultValue="5") float num2) {
         return new Subtraction(num1, num2);
     }
 	
 	@RequestMapping("/multiply")
-    public Multiplication multiplication(@RequestParam(value="num1", defaultValue="5") int num1, @RequestParam(value="num2", defaultValue="5") int num2) {
+    public Multiplication multiplication(@RequestParam(value="num1", defaultValue="5") float num1, @RequestParam(value="num2", defaultValue="5") float num2) {
         return new Multiplication(num1, num2);
     }
 	
 	@RequestMapping("/divide")
-    public Division division(@RequestParam(value="num1", defaultValue="5") int num1, @RequestParam(value="num2", defaultValue="5") int num2) {
+    public Division division(@RequestParam(value="num1", defaultValue="5") float num1, @RequestParam(value="num2", defaultValue="5") float num2) {
         return new Division(num1, num2);
     }
